move sample settings and index for running tests from cli into tests subdirectory, use simple collection for project servers as well so that the settings are less redundant
