+++ tests/pre-run.js
@@ -1,0 +1,5 @@
+require('coffee-script/register');
+require('babel-register'); // For using ES7 syntax
+require('babel-polyfill');
+
+require('./run');
