+++ tests/tutor/exercise-editing-and-qa.coffee
@@ -2,12 +2,12 @@ Helpers = require 'openstax-tutor/test-integration/helpers'
 {expect} = require 'chai'
 USERS = require './users.json'
 
-UserCollection = require '../helpers/users'
+SimpleCollection = require '../helpers/simple-collection'
 
-users = new UserCollection(USERS)
+users = new SimpleCollection(USERS)
 
-tutorTeacher = users.get('teacher', 'tutor')
-contentAnalyst = users.get('content-analyst')
+tutorTeacher = users.find('teacher', 'tutor')
+contentAnalyst = users.find('content-analyst')
 
 cases =
   '7651':
